chromeos: servod.conf: Add support for debug flag.
This commit allows a user to enable the debug flag for servod in the
context of the upstart job.  Simply start the job with DEBUG set to 1.

For example,

  start servod BOARD=samus DEBUG=1

BUG=chromium:580684
BRANCH=None
TEST=cros_workon hdctools, build packages and build an image for
beaglebone_servo.  Log into the beaglebone and start servod with the
debug option set to 1.  Verify that servod debug logs are generated.

Change-Id: Ib00821ea2314961f22505784928040ba8f2af0a4
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/323373
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>