servo: Add control to dynamically change loglevel.
It'd be nice to change the loglevel of servod without having to
restart the process entirely.  This change adds a control, loglevel,
which allows the user to query and set the log level without
restarting servod.

BUG=chromium:580684
BRANCH=None
TEST=cros_workon hdctools for the host.  Emerge hdctools, run servod
and use dut-control to query the current log level.  Verify that it is
'info'.  Use dut-control to change the loglevel to debug and verify
that the log messages appear with the proper debug formatting.  Use
dut-control again to change the loglevel back to info and verify that
no debug log messages are emitted.

Change-Id: I96c3da2bc14e88bd9df0c77a55dc93197cc515e5
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/323542
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>