servo: fix lucid overlay for EC-3PO
On lucid, the cpu_uart and the ec_uart are the same. This change sets
the cpu_uart_pty to the interface that is provided by EC-3PO and fixes
raw_ec_uart_pty.

BUG=chrome-os-partner:46054
BRANCH=None
TEST=emerge hdctools and start servod for lucid. Check that cpu_uart_pty
and ec_uart_pty are the same and raw_ec_uart_pty is the same as
uart2_pty.

Change-Id: Ieae4aa342ed4f224b378afeeb812d543caa44986
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326906
Commit-Ready: Anatol Pomazau <anatol@google.com>
Tested-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>