servod: Change EC-3PO PTYs to be owned by user.
Since the PTY permissions of the PTYs served by EC-3PO are 660, users
often have to add themselves to the 'tty' group, or use sudo to open
their serial consoles.  This isn't always desirable due to the fact
that some users might have customizations or other settings that they
want to use.

In order to make opening the PTYs be "seamless" as in the past, this
commit changes the owner and group of the PTYs to be that of the
user's which started servod by means of the sudo environment
variables.

Additionally, there was a change to inherit from the Uart class
instead of the FTDI Fuart class.

BUG=chromium:583528
BRANCH=None
TEST=cros_workon start hdctools; sudo emerge hdctools; Start servod
and check that the owner of the 'ec_uart_pty' is set to the user which
started servod.

Change-Id: If13f0a2f0f22a29c831cdf7e49c4a3efee424d63
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/330416
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>