servo: Don't raise an exception when atmega has no response
On the servo v3 with old atmega keyboard map, this exception is
raised. It makes servod failing to start and also blocks the update
test firmware_FlashServoKeyboardMap to flash a new keyboard map.
It is a deadlock.

Disable this exception in order to make to flash a new keyboard map.

BUG=chromium:551926
TEST=Ran servod on a servo v3 with old keyboard map.

Change-Id: I450f6745522db8ea3dd9f04170f447fa67ef0ba0
Reviewed-on: https://chromium-review.googlesource.com/332230
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>