servo: sentry: Reuse JTAG pins for PD MCU UART.
Sentry is a skylake board and in order to talk to the PD MCU, the JTAG
pins have been repurposed as the UART pins for interface 1.

Without this change, the usbpd_uart_pty will fail to initialize.

BUG=chrome-os-partner:51190
BRANCH=None
TEST=cros_workon start hdctools; sudo emerge hdctools; sudo servod -b
sentry and verify that servod starts successfully.

Change-Id: Ic9717592f37f38e60a561609219df15f74421b6e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/332841
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>