servo: Add advice to PD MCU UART bad interface err
Occasionally, an error in starting servo will happen because the
interface lists for a board is incorrect.  This fix is to override the
'dummy' interface with the 'ftdi_uart' interface for the index that is
being overridden.

This commit just adds a log message telling the user what might be
wrong.

BUG=chrome-os-partner:51190
BRANCH=None
TEST=cros_workon start hdctools; sudo emerge hdctools; Don't override
the interface in servo_interfaces.py, run servod and verify that advice
message is printed.

Change-Id: I5161fcfacdbe6e5e7c71d44bbe54281017a99811
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/332842
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>