servod: download_image_to_usb only downloads new images.
Servod will download whatever you tell it to download even if the last
image you told it to download already exists on the usb device.  Add in
a quick check to see if the last image path matches the current image
path to download and skip the transfer if it's the same saving time and
network bandwidth.

BUG=chromium:621994
TEST=Locally validated that servod will skip transfer on the same
image_path.

Change-Id: Ie5f14b7d2a207bf629271ba70d9308161a35eae8
Reviewed-on: https://chromium-review.googlesource.com/354610
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>