cros_ec_pd_softrec_power: Remove race condition.
In order to enter recovery mode, _power_on_bytype() relied upon
warm_reset keeping the AP in reset.  However, it was found that for some
platforms (samus, chell), warm_reset did not actually hold the AP in
reset.  This caused a race condition where if the AP booted up far
enough in normal mode, it would send a message to the PD MCU to jump to
its RW image.  This would happen before the EC had a chance to reboot
and hold the AP in reset. Failing to do so, when the request to boot
into recovery mode arrived, the AP would notice that the PD MCU was not
in its RO image. It would then send a message to the PD MCU to hard
reboot so that it would boot in its RO image.  However, the EC would
notice that the PD MCU was now in its RO image when it was previously in
its RW image.  This caused the EC to panic because it assumed that the
PD MCU had crashed.  When the EC panicked, it reset the entire system
and the recovery request was forgotten.

In order to reliably enter recovery mode, _power_on_bytype() now does the
following:

 - Hold PD MCU in reset.
 - Assert and deassert cold_reset.
 - Reboot the EC and have the EC now hold the AP in reset.
 - Release PD MCU reset.
 - Enter recovery mode.
 - Power down AP.
 - Enter recovery mode again.

By holding the PD MCU in reset until we've rebooted the EC keeping the
AP in reset, the PD MCU will never get the message to jump to its RW
image.  A side-effect of this however, is that in our first recovery
boot, the recovery reason will be that software sync failed to check the
version.  That's why we perform two recovery boots such that the
recovery reason will be "button press" once the function completes.

For platforms, regardless if they have a PD MCU or not, that have a
warm_reset control that actually holds the AP in reset, they should use
the cros_ec_softrec_power driver instead of this one.

BUG=chromium:606062
BUG=chromium:617832
BRANCH=None
TEST=platform_ServoPowerStateController passed for chell & samus on both
servo v2 and v3.

CQ-DEPEND=CL:355505

Change-Id: I4a577ec788a54c8efffa72450586ee95e194251e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/355580
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>