servo_v4: Update default interfaces to work with servo micro.
To prepare for the work to enable servo v4 and servo micro to be served
by the same servod process, I'll be updating the servo v4 interface to
allow servo micro to use the first 10 servo interfaces so that no
changes need to be done to the existing board interfaces.

Also delete the existing servod._init_dummy() method since it's not
needed anymore since I look for dummy interfaces in the interface init
loop and deal with them explicitly.

BUG=chromium:599538
TEST=start servod and can still access servo console

Change-Id: I984d23f7e8db3a46c7205952fd535ffd742f7465
Reviewed-on: https://chromium-review.googlesource.com/358881
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>