start_servod.conf: Update the pre-start to keep the old config.
This is to ensure backwards compatibility if the need to downgrade comes
up.  Prior, if we downgraded after an upgrade, servod would not start up
since the config file from 'config' to 'config_9999'.  This change keeps
the old config file as well as creating the new style config file.

BUG=None
TEST=on a lab beaglebone, upgraded then downgraded and ensured the downgraded
beaglebone was able to start up servod.

Change-Id: I232b512b7edc7290d04ab035aa1f7b85e067ad74
Reviewed-on: https://chromium-review.googlesource.com/362691
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@google.com>