servo: asuka: Reuse JTAG pins for PD MCU UART.
Asuka is a skylake board and in order to talk to the PD MCU, the JTAG
pins have been repurposed as the UART pins for interface 1.

Without this change, the usbpd_uart_pty will fail to initialize.

BUG=None
BRANCH=None
TEST=cros_workon start hdctools; sudo emerge hdctools; sudo servod -b
asuka and verify that servod starts successfully.

Change-Id: I406a4656c45bd0394b2dbd0743f9393a63a10c3e
Reviewed-on: https://chromium-review.googlesource.com/366540
Commit-Ready: Ted Kuo <tedkuo@ami.com.tw>
Tested-by: Ted Kuo <tedkuo@ami.com.tw>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>