servo v2: Set dut_hub_pwren init to be on.
In autotest during servo init, the dut_hub_pwren control is explicity
set to 'on' and never touched afterwards.  In servo v3, dut_hub_pwren
control already has the init set to 'on' and in servo v4, there does not
exist a dut_hub_pwren signal (the closest is dut_usb_hub_reset).  To
avoid the need to create a fake dut_hub_pwren for servo v4, let's have
the init set it to 'on' for servo v2 and take out the code in autotest
that toggles dut_hub_pwren.

BUG=None
TEST=Test that firmware_FAFTSetup still works on a servo v2 and gnawty.

Change-Id: I2f94c7172a80578823254f5fd486aa7104f35387
Reviewed-on: https://chromium-review.googlesource.com/383672
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>