servod: Update safe usb probing and add 2 additional RPCs.
Added the following rpcs to enable safe usb probing.
- safe_switch_usbkey_power
- safe_switch_usbkey

Autotest toggles the power and mux direction for the usb key prior to
probing for it.  I added a safe option for them so that the
power/direction can be toggled without screwing with another servod's
probing.

I also refactored the blocking to be a bit more foolproof and easier to
use.

BUG=chromium:599538
TEST=locally on a servo host with 3 servo v4s probing simultaneously and
all found their usb devices.

Change-Id: I4e35f102286015856b58bb0c6c6174c85e201120
Reviewed-on: https://chromium-review.googlesource.com/391236
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>