servod: Add in a probe_timeout arg to the download_image_to_usb RPC.
During deployment, the image dut step downloads the image to the usb, if
there are more than 3 duts trying to do so, the duts over 3 will timeout
trying to probe for a usb device and fail.  Deployments will
generally have more than 3 duts so give the option to disable the timeout
for the probe call.

BUG=chromium:599533
TEST=on servo host with 8 servo v4 trying to image the drive, none of
them timed out when timeout was disabled.

Change-Id: I0da0689815fdd09534c332dac15ac7ae86eb1f7a
Reviewed-on: https://chromium-review.googlesource.com/398282
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>