servo_glados_overlay: Add in servo_{v4,micro}_interface to raw_usbpd_uart_pty.
servo_glados_overlay defines raw_usbpd_uart_pty on an interface
specifically for servo v2/v3 (that is reworked) that is not compatible
with servo v4/uServo.  This is a sticky situation since this control
really should belong in the servo v2/v3 overlay but at the same time
doesnt since those controls don't really exist without a rework on the
v2/v3.

A proper solution is still being hashed out but this is blocking servod
from starting which is blocking all boards that use this overlay from
deploying.  Right now just define a servo v4/uServo specific interface
for this control so that servod can get going.  It's expected that usbpd
won't work this way but we're not using it as part of deployments and the
majority of testing doesn't as well so we're ok with that.

BUG=chromium:660922
TEST=Started servod with change with servo v4/uServo and asuka and did
not see the conflict.

Change-Id: If444c18460103df0b4d921163f8445424d398f1c
Reviewed-on: https://chromium-review.googlesource.com/405757
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>