servod: Don't die if remap control doesn't exist.
In the servo_x86-alex_overlay, fw_up remaps rec_mode which assumes
fw_up exists already in the servo overlay (which is not the case in
servo v4).  If the control doesn't exist in the first place, no need to
die trying to remap it.

BUG=None
TEST=start up servod with servo v4 and servo_x86-alex_overlay
successfully.

Change-Id: Ia2834c36c5e57a8fc9a1c78e03c729b3df226693
Reviewed-on: https://chromium-review.googlesource.com/406631
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>