cros_ec_softrec: Allow time to reach rec screen.
It takes some time to reach the recovery screen for devices.  Most of
the time it's fairly quick, where other boards make take a longer time
to do so.  The power state driver would immediately yield after issuing
"power_state:rec" but it was not waiting until the device had actually
booted up into recovery.

This commit adds a parameter called "boot_to_rec_screen_delay" which can
be used in the board's XML file.  It will default to 5 seconds, but any
board can override that if they require a longer delay to reach the
recovery screen.

BUG=chromium:666872
BRANCH=None
TEST=run platform_ServoPowerStateControllerUSBUnplugged for kevin on
servo v2 and verify that it passes.
TEST=set boot_to_rec_screen_delay to 10s for reef and run
platform_ServoPowerStateControllerUSBPluggedin for reef on servo v2 and
verify that it passes.

Change-Id: I757906271a373ea7c570f5aa84c96b51c341a732
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/412975
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>